(maint) Make external CA acceptance test work with SELinux (TRY #2)
The previous attempt to make Apache and Passenger play nicely with
SELinux was sufficient to get httpd started, but insufficient to allow
Passenger to run.  Passenger switches users and creates sockets all over
the place, which SELinux denies.

This patch simply disables SELinux for the duration of the external CA
testing and re-enables it if it was originally enabled.