(#20128) Handle event dequeues when target resource has no events registered
In the case where a resource triggers an event with invalidate_refreshes
enabled (i.e. a service starts), the dequeue process would attempt to
invalidate other events registered for this resource.  If no other
notifications had been received on the resource already, then the hash element
for the resource would be nil.  The resource would then fail:

  Error: /Stage[main]//Service[service]: Could not evaluate: undefined method `[]=' for nil:NilClass