(maint) Ensure results hash is defined for module install
If the puppet module tool attempted to install a module from the
filesystem that does not exist, it would raise an error before a
variable was defined that the ensure clause needed to modify. Doing so
would raise a NoMethodError on nil, which masked the underlying issue.
This commit defines the variable at the beginning of the method to
ensure that the ensure clause works as expected.