(maint) Update COMMITTERS with stable and master branch strategy
Without this patch the current branching strategy of creating unique
branches for each minor release requires CI systems be reconfigured for
each new minor release.  This is a problem because it is time consuming,
error prone, and expensive to reconfigure the CI system for each minor
release.  This, in turn, creates a road block to releasing more often.

This patch addresses the problem by documenting and prescribing a new
branching strategy for committers.  The strategy results in branch names
that do not change over time.  The `stable` branch always refers to the
current release which is receiving bug fixes.  The `master` branch
always refers to the next release which is receiving new functionality.

The release branches are no longer direct descendants (fast forwards) of
one another.  Instead, they are independent histories as in the git flow
model.  The security branch is special and receives change sets which
are then rebased and merged into specific branches for the specific
release versions which are affected by the vulnerability.

[ci skip]