(Maint) Properly mark ruby 2 test to not run without sqlite
It seems that rspec example groups do not compose their pending
conditions. For instance if and outer describe is

  describe "foo", :if => we_have_the_power? do
  ...

and a context is inside that with

  context "fighters", :if => band_members_available? do
  ...

then the context will still be executed if the band members are
available but we don't have the power.

This was causing the ruby 2 test for active record to execute even if we
didn't have active record available.