(Maint) Remove disused puppet.gemspec
The puppet.gemspec that sat at the top level of the puppet project was
not needed for anything other than the version information when using
bundler with puppet when doing development. The gemspec actually caused
a warning to be issued about duplicate definition of PUPPETVERSION
because of the way the gemspec required the puppet/version file.

This updates the Gemfile to force a fake version for the locally used
code and removes the gemspec which isn't needed.