(maint) Fix puppet module build empty tar issue on Solaris
Without this patch `puppet module build` produces an empty tar file when
run on Solaris.  This is obviously a problem.  The root cause is that
tar defaults to writing the archive to the tape device instead of STDOUT
as was expected.

This patch fixes the problem by explicitly using `tar cf - ...` to write
the archive to STDOUT.

Paired-with: Andrew Parker <andy@puppetlabs.com>