(#20303) Don't rely on automatic array conversion
Previously, we were unwittingly relying on ruby's automatic conversion
from an array to a string, which fails in ruby 1.9. This was noticed
when running `puppet resource user` as it reported all users were
absent.

Now we explicitly convert from an array to a string.

This commit doesn't have any tests, because the existing acceptance
tests catch this when run under 1.9