(#19886) Add a function that performs heredoc syntax checking for JSON
By using a Heredoc syntax string of "json", or that ends with "...+json"
the function added in this commit will perform a test parse and report
the failure.

The function can also be called directly from a manifest to syntax check
an aribtrary string - e.g. check_json_syntax("am I valid", "json").

When used in a heredoc the same is triggered by specifying the syntax in
the heredoc tag e.g. $a = @(END:json) ...