(#19886) Add ability to specify escapes for dq strings
It was a nuisance to not be able to control the escapes in a dq string.
This adds the ability to specify the escapes for a dq string the same
was as for a sq string.

In addition, the notation '/' turns on all escapes. The default for a dq
strings is not also to not use any escapes e.g. @(END) means escapes are
verbatim.