Revert "(Maint) Remove disused puppet.gemspec"
This reverts commit 588b874f2fc42cf1b9d21016aa36d98288df2479.

This commit is being reverted because the change to remove the gemspec
was based on incorrect information.  According to the commit message,
the gemspec file was removed for two reasons.  First, it was thought the
gemspec was not being used for anything other than local development of
Puppet.  Second, the gemspec produced a warning about redefining the
PUPPETVERSION constant.

The gemspec is actually being used in our automated testing of other
projects, specifically stdlib.  The standard library testing matrix uses
the gemspec located in Puppet to satisfy the dependency of stdlib on
Puppet.  The gemspec allows the `stable` and `master` branch to be used
in addition to the officially released Puppet gem on rubygems.org.

Second, the warning is a not a result of the gemspec itself but rather a
result of the way the gemspec is loading 'puppet/version'  We can avoid
the warning while also preserving the gemspec.