(maint) move puppet.gemspec to .gemspec
Without this patch some users assumed the puppet.gemspec is intended to
build the Puppet Gem.  This is a problem because the packaging rake
tasks are what actually build the official Gem.

This patch addresses the problem by moving the gemspec to a hidden file
so that other people are less likely to be confused by the presence of
the gemspec file.