(maint) Fix PUPPETVERSION warning from gemspec
Without this patch Ruby issues a warning about the redefinition of the
PUPPETVERSION constant.  This warning is caused by the gemspec using a
fully qualified require statement instead of a relative $LOAD_PATH based
require.

This patch fixes the problem by catching the LoadError and adding
Puppet's `lib` directory to the head of the $LOAD_PATH.  This change
causes the `require` statement to correctly load puppet/version only
once, which avoids the redefined constant message.

In addition, a comment about the intended use of the gemspec has been
added to help avoid confusion.  This gemspec is intended to allow third
party code to automatically satisfy their dependency on Puppet using
Bundler.  The gemspec is not intended to actually build the package.

Paired-with: Adrien Thebo <adrien@puppetlabs.com>