(maint) Do not use a fake hiera version
Without this patch hiera master cannot be used from Git source to
execute in Bundler.  This is because Puppet incorrectly specifies a
dependency on Hiera 99.0.0 when the correct version of Hiera master is
currently "1.2.1"

This patch addresses the problem by simply removing the fake version,
allowing the correct version to satisfy the dependency.