(maint) Make sqlite3 use an in-memory DB for specs
Without this patch the spec tests are configuring sqlite in a manner
that writes the database to a file named ':memory'.  This is a problem
because filesystem databases are slower than in-memory databases and the
tests are not cleaning up this file after completing.

The change in 60e78b2d36f23877dc7b50e72578ad123d62c1f0 caused this
change in behavior, likely by accident, because the database connection
string was ':memory:' and the patch changed the string to ':memory'
likely because this looks like a Ruby symbol.

This patch addresses the problem by restoring the original connection
string of ':memory:'