(#20308) Fix dropping paths from fileset results
This fixes the issue in #20308 where some paths get dropped from the fileset
results. When a directory contained both subdirectories and files, and the
last entry was a file, it could cause the recursion to exit early. This
removes the complexity of tracking the next recursion level separately.

Paired-with: Andrew Parker <andy@puppetlabs.com>