(maint) Handle different ssh service names in `puppet resource service` test
This test was using a regex to match the output for the ssh/sshd service
from puppet resource, but was explicitly invoking the command with the
name ssh. Since we can check the state of that service from the list of
all instances (in which case the regex is actually beneficial), there's
no need to look at the specific service separately. This combines the
two steps, and makes the test more robust to different names for the ssh
service.