(maint) Fix install module acceptance test on precise
Without this patch the `with_modulepath.rb` acceptance test is failing
on Ubuntu 12.04.  The cause of this problem is that `puppet apply` is
being used to purge the `/etc/puppet/modules2` directory, which doesn't
appear to be working as expected.  This itself is concerning, but the
concern is unrelated to the expected behavior expressed in the test.

This patch addresses the problem by simply using `rm -rf` to purge the
acceptance test instead of using `puppet apply`.