(maint) Fix enc_provides_node_when_storeconfigs_enabled on precise
Without this patch the enc_provides_node_when_storeconfigs_enabled is
failing on Ubuntu 12.04 LTS with MRI 1.9.3-p0.  The root cause of the
failure is that the ENC script used does not output valid YAML data,
it's lacking the `--- ` header.

This patch addresses the problem by using `Hash#to_yaml` to produce
valid YAML output for the ENC API.