(Maint) Inject Pidlock dep into Daemon
While trying to understand several different code paths in the
Puppet::Daemon code the realization that the tests for the daemon code
had to mock internal methods on the class under test caused me worry. By
removing the mocking of the internal methods it was obvious that there
was another collaborator that is a locking service at play. Services
should normally be injected into, and not created by, the consumer of
the service.

This injects the Puppet::Util::Pidlock into the Daemon and reworks the
points of construction of Daemon objects so that the correct information
is available at the point of construction. This require moving the
Daemon creation from the preinit methods to the setup or run methods in
order to ensure that command line processing and therefore Puppet#[] was
ready at the point of construction so that the correct Puppet[:pidfile]
could be used.