(Maint) Remove unused register code for Server
The Puppet::Network::Server class seems to have atrophied over time.
There are multiple redundant or unused code paths in it.

This commit removes the "registration" facility for the network server
to handle certain indirections. There were no users of these outside of
the tests and there is no functionality that they provided. The only
thing the registration of indirection names did was check to see if such
an indirection existed. I believe that at one point these had been meant
to act as a per-server handling facility (as indicated by one of the
test names), but they no longer serve such a purpose.