(maint) Add Fedora Core suitability check for password redaction
Without this patch the Fedora Core 18 system, prefixed by `fc-` is
assumed to be able to manage password.  This is a problem because the
system does not have the shadow gem installed by default.

This patch adds `fc-` to the list of platforms checked for suitability
before running the acceptance test.