(maint) Remove acceptance test for puppet --parseonly deprecation
Without this patch the acceptance test for the puppet master --parseonly
and puppet agent --parseonly is failing.  This is a problem because in
Puppet 3 the deprecation has been removed and instead Puppet reports an
error that the command line is incorrect, 1 argument was provided
instead of the expected 0 number of arguments.

This patch addresses the problem by changing the acceptance test to no
longer validate the presence of the deprecation warning and instead
validate the expected behavior of `puppet parser validate`.