(maint) `--parseonly` lambda arity must match number of arguments
Running `puppet apply --parseonly` and `puppet master --parseonly` fails on
ruby 1.9, because the arity of the option method was 0, yet was being invoked
with an argument, the manifest to parse. Under ruby 1.8, this mismatch was
allowed, but not 1.9.

This patch builds upon Josh's patch in e9c25f6, the previous patch only
addressed `puppet apply` while the acceptance test uses both `puppet apply` and
`puppet master`.  In order to address the test failure this patch adds the
block arguments for both apply and master.

See acceptance/tests/ticket_6928_puppet_master_parse_fails.rb