(#14615) Exclude solaris from service name acceptance test
Without this patch the ssh service acceptance tests that ensures the
service providers do not modify the system in d4ced36 are failing on
Solaris 10.  The failures are a result of the service name for ssh being
wildly different on solaris.  On solaris the name is
`svc:/network/ssh:default` which differs considerably from the Linux
convention of `ssh`.

This patch addresses the problem by excluding the acceptance test from
running on Solaris systems.