(maint) Corrects manual edit in generated eparser.rb
NEVER CHANGE A GENERATED FILE! There is a big warning at the top
of the file!

In this case, the puppet/util/loadedfile.rb was removed and the
egrammar.ra file contained a require of this file. A manual edit was
made of the generated egrammar.rb where the require was removed. This
means that on the next regeneration of the grammar this require
reappears.

This corrects this mistake.