(maint) Use activerecord 2.2.3 on CentOS 5
Without this patch the ENC acceptance test is attempting to install
activerecord 3.0.20 on CentOS 5.  This is a problem because activerecord
3.0.20 requires a more recent version of rubygems than is available in
CentOS 5 resulting in activerecord not being installed.

This patch addresses the problem by using a selector to install an older
version of activerecord on EL 5 systems, allowing the test to pass
successfully.