(#19760) Use combine when running pkginfo
pkginfo was previously executed with combine => false so stderr was
discarded. This way puppet was unable to detect any errors the pkginfo
might print, especially the message `ERROR: information for
"some_package" was not found`.

As a result the provider failed every time it tries to check the current
state of an absent package.