Maint: Improve documentation of filebucket attributes
The docs were unclear and partially inaccurate about the behavior of the
filebucket type's path and server attributes. I have tested the precedence of
these and documented the actual behavior. This commit also improves docs for the
file type's "backup" attribute and the static compiler, in light of what was
learned about default values.