(PE-158) speed license check
Prior to this commit, the CertificateAuthority did not expose any
methods that allow Puppet Enterprise license checking to not require
reading the certificate files twice (once for `list` and again for
`verify`).  This commit allows the CertificateAuthority to list
certificates as objects and then, in turn, be able to issue a judgement
of whether or not these certificate objects are valid/live.