(maint) stub `dpkgquery` in aptitude specs
The Puppet::Provider.command method will attempt to look up a command
and will verify that it's present on the system with Puppet::Util#which.
If the underlying command is not present an exception will be raised,
and this was causing spec tests to fail. This commit stubs out the
`dpkgquery` command in the aptitude spec provider so that the specs can
be run on platforms that don't have /usr/bin/dpkg-query.

Paired-with: Joshua Partlow <joshua.partlow@puppetlabs.com>