(#20726) Include password for useradd managehome tests on windows
The new acceptence tests for managing the home directory fails on
Windows because windows requires a sufficiently complex password to be
specified when creating the home directory.

With that said, managing the home directory is broken on windows with
ruby 1.9 (#20768). This marks the test as pending when not running ruby
1.8.

Paired-with: Josh Cooper <josh@puppetlabs.com>