(#20726) Manage home has issues on Solaris
On Solaris, executing `usermod -d dir username` where dir is the user's
home directory as reported by `puppet resource user username` will fail
with the error:

    UX: useradd: ERROR: Path must not start with '/home/'

It appears the new directory must be of the form
'/export/home/username'. Since this issue is unrelated to #20726, I'm
marking this test as pending on Solaris.