(#20768) Use the correct Win32API method signatures
In ruby 1.8, we could call CloseHandle passing it the handle (an opaque
unsigned integer value), and due to the 'P' in the method signature,
automatically arrayify and pack the handle. This wasn't done intentionally,
it just happened to work.

But ruby 1.9 no longer automatically coerces the parameter, instead
the CloseHandle method (implemented by the windows-pr gem) tries to pack
the handle, generating an error `can't convert Fixnum into String`. This
prevents puppet from managing a user's password and home directory.

The proper fix is to define the CloseHandle method signature correctly, so
that the windows-pr layer doesn't attempt to coerce the handle value.

The UnloadUserProfile API had a similar problem, it takes the handle as an
'L' input paramter (instead of a pointer).