(#11108) Test coverage for remote pkg file
The current implementation of tests for the pkgdmg provider only test
the behavior for packages without a file extension if they're being
installed locally. This commit adds tests for remote pkgs so that if a
file extension is not given in the package tile, it's extracted from the
package source.