(#20742) Only use certifcate info if there is a CN
Because of the changes to puppet/util/ssl for this issue, the DN
was now always able to be parsed into something. This uncovered that the
logic around deciding on the node name and authentication status for the
request was incorrectly based off of simply the existance of a DN in the
headers. This caused a DN that was missing a CN to have an unknown node
for a request, which could be marked as authenticated if the verify
header was "SUCCESS".

This patch fixes it by basing the decision of using the SSL headers on
the existance of the CN instead of the DN. Now a CN must be present
(e.g. a valid DN) for the verify header to be used. When no CN is
present, then it is treated the same as no DN being present, which
causes a reverse DNS lookup to be done for the node name.